Split triple store quering in a new file, resolves #57.
